Move production traits to Production folder + fixes in Production.cs
